requirements.txt
python-dotenv
langchain
langchain_core
youtube-transcript-api
youtube-transcript-api==0.1.6
streamlit 
langchain-google-genai
langchain-community
google-generativeai
faiss-cpu
streamlit 